refactor: modal theme
