--- src/common/tools/chakra-ui/components/Modal.ts
@@ -1,46 +1,0 @@
-import { modalAnatomy as parts } from '@chakra-ui/anatomy';
-import { createMultiStyleConfigHelpers } from '@chakra-ui/styled-system';
-
-const { definePartsStyle, defineMultiStyleConfig } = createMultiStyleConfigHelpers(parts.keys);
-
-type defaultProps = ReturnType<typeof defineMultiStyleConfig>['defaultProps'];
-
-const modalDefaultProps = {
-  isCentered: true,
-} as defaultProps; // type casting is necessary because theme is not accepting ModalProps
-
-export const Modal = defineMultiStyleConfig({
-  baseStyle: definePartsStyle({
-    header: {
-      color: 'gray.900',
-      fontWeight: '500',
-      letterSpacing: '-0.02em',
-      fontSize: '36px',
-      p: 0,
-    },
-    dialog: {
-      bg: 'border.box',
-      border: '2px solid transparent',
-      borderRadius: '3xl',
-      boxShadow:
-        '0px 100px 80px rgba(1, 133, 195, 0.05), 0px 12.5216px 10.0172px rgba(1, 133, 195, 0.025)',
-      mx: { base: 6, md: 0 },
-      p: '16',
-      pb: '8',
-      position: 'relative',
-      textAlign: 'center',
-    },
-    overlay: {
-      bgColor: 'rgba(148, 163, 184, 0.1)',
-      backdropFilter: 'auto',
-      backdropBlur: '10px',
-    },
-    body: { p: 0, color: 'gray.600', mt: '4' },
-    footer: {
-      justifyContent: 'center',
-      p: 0,
-      mt: '6',
-    },
-  }),
-  defaultProps: modalDefaultProps,
-});
